requirement.txt
appdirs==1.4.3
asgiref==3.2.7
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==5.1
crypto==1.4.1
cryptography==2.4.2
distlib==0.3.0
Django==3.0.6
django-debug-toolbar==2.2
django-filter==2.3.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-mptt==0.11.0
django-pure-pagination==0.3.0
django-redis==4.12.1
djangorestframework==3.11.1
elasticsearch==2.4.1
fabric==2.5.0
Faker==4.0.3
filelock==3.0.12

idna==2.8
importlib-metadata==1.6.0
importlib-resources==1.5.0
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.10
Markdown==3.2.1
MarkupSafe==1.1.0
MouseInfo==0.1.3
Naked==0.1.31
netaddr==0.8.0
numpy==1.19.1
opencv-python==4.4.0.40
paramiko==2.4.2
pathtools==0.1.2
Pillow==5.4.1
pipenv==2018.11.26
psutil==5.4.8
pyasn1==0.4.5
PyAutoGUI==0.9.50
pycparser==2.19
PyGetWindow==0.0.8
pymemcache==2.1.0
PyMsgBox==1.0.8
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyotp==2.3.0
pyperclip==1.8.0
PyQt5==5.15.0
PyQt5-sip==12.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-engineio==3.2.3
python-socketio==3.1.1
python3-memcached==1.51
PyTweening==1.0.3
pytz==2020.1
pywin32==224
PyYAML==5.1.1
raven==6.10.0
redis==3.5.3
requests==2.22.0
sentry-sdk==0.16.2
shellescape==3.4.1
six==1.12.0
some-package==0.1
SQLAlchemy==1.2.16
sqlparse==0.3.1
text-unidecode==1.3
unrar==0.3
urllib3==1.25.3
virtualenv==20.0.20
virtualenv-clone==0.5.4
watchdog==0.10.2
Werkzeug==0.14.1
winshell==0.6
xmltodict==0.11.0
zipp==3.1.0
